Replicate snap refresh timer fix for master
The fix in #212 should apply equally to both master and worker, so this
moves it to the base.
